renamed services.stub back to spec, to fix build errors (missing jasmine namespace)
