+++ projects/ui-framework/src/lib/form-elements/date-picker/date-range-picker/date-range-picker.component.spec.ts
@@ -24,7 +24,7 @@ import { DateInputDirectiveModule } from '../date-input-directive/dateinput.dire
 import {
   utilsServiceStub,
   mobileServiceStub,
-} from '../../../tests/services.stub';
+} from '../../../tests/services.stub.spec';
 import { InputEventType } from '../../form-elements.enum';
 import {
   MatDatepicker,
